pyproject.toml
[tool.poetry]
name = "ecommerce-data-analysis"
version = "0.1.0"
description = ""
authors = ["tabatha wiggers <tatchiwiggers@gmail.com>"]


[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
pytest = "^8.3.2"
httpx = "^0.27.2"
selenium = "^4.24.0"
bs4 = "^0.0.2"
pyvirtualdisplay = "^3.0"
undetected-chromedriver = "^3.5.5"
setuptools = "^74.1.2"
pydantic = "^2.9.1"
email-validator = "^2.2.0"
pydentic = "^0.0.1.dev3"
snowflake-connector-python = "^3.12.2"
python-dotenv = "^1.0.1"
snowflake-sqlalchemy = "^1.6.1"
apache-airflow = "^2.10.1"
dbt-core = "^1.8.6"
dbt-snowflake = "^1.8.3"
azure-cosmos = "^4.7.0"
pendulum = "^3.0.0"
pandas = "^2.2.3"
flake8 = "^7.1.1"
matplotlib = "^3.9.2"
streamlit = "^1.38.0"
seaborn = "^0.13.2"
xlsxwriter = "^3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
